+++ tslib/src/functions/check-marker/models.ts
